--- src/cleanup.ts
@@ -1,7 +1,0 @@
-import { cleanupDOM } from './makeRenderMethods';
-import { cleanupEnvVars } from './setEnvVars';
-
-export function cleanup() {
-  cleanupDOM();
-  cleanupEnvVars();
-}
