+++ src/index.ts
@@ -1,3 +1,2 @@
 export { default as getPage } from './getPage';
-export { initTestHelpers } from './testHelpers';
-export { cleanup } from './cleanup';
+export { initTestHelpers, cleanup } from './testHelpers';
