+++ src/router/RouterProvider.tsx
@@ -2,9 +2,9 @@ import React, { useState, useCallback, useRef } from 'react';
 import { RouterContext } from 'next/dist/next-server/lib/router-context';
 import { formatUrl } from 'next/dist/next-server/lib/router/utils/format-url';
 import makeRouterMock, { PushHandler } from './makeRouterMock';
-import { useMountedState } from './utils';
-import { ExtendedOptions, MakePageResult, PageObject } from './commonTypes';
-import { RuntimeEnvironment } from './constants';
+import { useMountedState } from '../utils';
+import { ExtendedOptions, MakePageResult, PageObject } from '../commonTypes';
+import { RuntimeEnvironment } from '../constants';
 
 export default function RouterProvider({
   pageObject,
